Split the fault handling from MemMap.cpp to MemFault.cpp.
