+++ Core/Core.vcxproj
@@ -129,8 +129,7 @@
   <ImportGroup Label="ExtensionSettings">
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
-  <Target Name="CreateTlogDir"
-          BeforeTargets="CustomBuildStep">
+  <Target Name="CreateTlogDir" BeforeTargets="CustomBuildStep">
     <MakeDir Directories="$(TLogLocation)" />
   </Target>
   <PropertyGroup>
@@ -389,6 +388,7 @@
     <ClCompile Include="HLE\sceUsbCam.cpp" />
     <ClCompile Include="HLE\sceUsbMic.cpp" />
     <ClCompile Include="HW\Camera.cpp" />
+    <ClCompile Include="MemFault.cpp" />
     <ClCompile Include="MIPS\IR\IRAsm.cpp" />
     <ClCompile Include="MIPS\IR\IRCompALU.cpp" />
     <ClCompile Include="MIPS\IR\IRCompBranch.cpp" />
@@ -918,6 +918,7 @@
     <ClInclude Include="HLE\sceUsbCam.h" />
     <ClInclude Include="HLE\sceUsbMic.h" />
     <ClInclude Include="HW\Camera.h" />
+    <ClInclude Include="MemFault.h" />
     <ClInclude Include="MIPS\IR\IRFrontend.h" />
     <ClInclude Include="MIPS\IR\IRInst.h" />
     <ClInclude Include="MIPS\IR\IRInterpreter.h" />
